libelement: renamed interpreter_object_model to object, and object to object_internal
