+++ libelement/src/object.cpp
@@ -1,11 +1,20 @@
 #include "element/common.h"
-#include "element/interpreter_object_model.h"
+#include "element/object.h"
 #include "object_model/compilation_context.hpp"
 #include "interpreter_internal.hpp"
 #include "instruction_tree/evaluator.hpp"
 #include "object_model/intermediaries/declaration_wrapper.hpp"
 
-element_result element_object_model_compile(
+element_result element_declaration_to_object(
+    const element_declaration* declaration,
+    element_object** output)
+{
+    auto wrapped_declaration = std::make_shared<const element::declaration_wrapper>(declaration->decl);
+    *output = new element_object{ std::move(wrapped_declaration) };
+    return ELEMENT_OK;
+}
+
+element_result element_object_compile(
     element_interpreter_ctx* context, 
     const element_object* compilable,
     element_object** output)
@@ -20,16 +29,7 @@ element_result element_object_model_compile(
     return result;
 }
 
-element_result element_declaration_to_object_model_declaration(
-    const element_declaration* declaration,
-    element_object** output)
-{
-    auto wrapped_declaration = std::make_shared<const element::declaration_wrapper>(declaration->decl);
-    *output = new element_object { std::move(wrapped_declaration) };
-    return ELEMENT_OK;
-}
-
-element_result element_object_model_call(
+element_result element_object_call(
     element_interpreter_ctx* context,
     const element_object* callable,
     const element_object arguments[],
@@ -51,7 +51,7 @@ element_result element_object_model_call(
     return result;
 }
 
-element_result element_object_model_index(
+element_result element_object_index(
     element_interpreter_ctx* context,
     const element_object* indexable,
     const char* index,
