+++ libelement/src/object_model/constraints/constraint.hpp
@@ -2,7 +2,7 @@
 
 //SELF
 #include "typeutil.hpp"
-#include "object_model/object.hpp"
+#include "object_model/object_internal.hpp"
 
 namespace element
 {
