+++ libelement/src/object_model/expressions/anonymous_block_expression.hpp
@@ -5,7 +5,7 @@
 
 //SELF
 #include "expression.hpp"
-#include "object_model/object.hpp"
+#include "object_model/object_internal.hpp"
 
 namespace element
 {
