+++ libelement/src/object_model/expressions/expression_chain.hpp
@@ -2,7 +2,7 @@
 
 //SELF
 #include "expression.hpp"
-#include "object_model/object.hpp"
+#include "object_model/object_internal.hpp"
 #include "object_model/capture_stack.hpp"
 
 namespace element
