+++ libelement/src/object_model/intermediaries/anonymous_block_instance.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 //SELF
-#include "object_model/object.hpp"
+#include "object_model/object_internal.hpp"
 #include "object_model/capture_stack.hpp"
 
 namespace element
